# src/main/java/microarch.delivery/adapters/in/http/api/CompleteOrderApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package microarch.delivery.adapters.in.http.api;

import microarch.delivery.adapters.in.http.model.Error;
import java.util.UUID;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-24T15:35:30.575475+03:00[Europe/Moscow]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "CompleteOrder", description = "the CompleteOrder API")
public interface CompleteOrderApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_COMPLETE_ORDER = "/api/v1/couriers/{courierId}/orders/{orderId}/complete";
    /**
     * POST /api/v1/couriers/{courierId}/orders/{orderId}/complete : Завершить заказ
     * Позволяет завершить заказ
     *
     * @param courierId Идентификатор курьера (required)
     * @param orderId Идентификатор заказа (required)
     * @return Успешный ответ (status code 200)
     *         or Некорректные параметры запроса (status code 400)
     *         or Конфликт при завершении заказа (status code 409)
     *         or Внутренняя ошибка сервиса (status code 500)
     */
    @Operation(
        operationId = "completeOrder",
        summary = "Завершить заказ",
        description = "Позволяет завершить заказ",
        tags = { "CompleteOrder" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Успешный ответ"),
            @ApiResponse(responseCode = "400", description = "Некорректные параметры запроса", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "409", description = "Конфликт при завершении заказа", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервиса", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = CompleteOrderApi.PATH_COMPLETE_ORDER,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> completeOrder(
        @Parameter(name = "courierId", description = "Идентификатор курьера", required = true, in = ParameterIn.PATH) @PathVariable("courierId") UUID courierId,
        @Parameter(name = "orderId", description = "Идентификатор заказа", required = true, in = ParameterIn.PATH) @PathVariable("orderId") UUID orderId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 0, \"message\" : \"message\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}